Allow inline html in _index.Rmarkdown
Starting with hugo 0.60, the default markdown renderer changed from
blackdown to goldmark. By default, goldmark does not render raw HTML so
specifying raw html in _index.Rmarkdown (in this instance to customize
an image's display) created an html file with <!-- raw HTML omitted -->
rather than the correct html to display the image.

You can make goldmark include raw html by switching it to `unsafe` mode
in the `config.toml` which this commit does.

As an FYI, An alternative is to mv `_index.Rmarkdown` to `_index.Rmd`
which will switch the markdown engine from goldmark to pandoc which
still renders raw html.